internal-fix(evaluation): round-four polish on the decode extraction
Tightens the NaN pin docstring to the two-line cap, narrows the loud-
failure claim to categorical truncation and pins the silent tail-
truncation case beside it, softens the [0, 1] return claim to the
convention it is, aligns the target path's dtype with the pred path via
.float(), and notes where exact-value coverage lives.

+++ src/synth_setter/evaluation/predict_vst_audio.py
@@ -196,7 +196,8 @@ def main(
 
             out_target = os.path.join(sample_dir, "target.wav")
             if rerender_target and target_params is not None:
-                target_params_ = target_params[j].numpy()
+                # .float() aligns the target path with the pred path's float32 contract.
+                target_params_ = target_params[j].float().numpy()
                 target_synth_params, target_note_params = decode_model_output(target_params_, spec)
 
                 new_target = render_params(
